requirement.txt
python-dotenv
numpy 
pandas
scikit-learn
dill
pymongo[srv]==3.6
pymongo
certifi
pyaml
